1.2.0 - weaken http provider dependency
Previous versions required an apache-airflow-providers-http 1.x version,
but this is overly restrictive - we work with all known versions of this
package. Loosen the constraint from "^1.1.1" to ">=1.1.1".

Also increase python version to 3.9 or later and fix a poetry bug with
docutils.

Tested via latest airflow in `docker compose`. This package's usage of
`HttpHook` is extremely simple and should work broadly.